Return connection in env var for flow
