+++ src/promptflow/tests/test_configs/flows/flow_with_connection_ref_environment_variables/flow.dag.yaml
@@ -1,0 +1,17 @@
+environment_variables:
+  API_KEY: ${azure_open_ai_connection.api_key}
+inputs:
+  key:
+    type: string
+outputs:
+  output:
+    type: string
+    reference: ${print_env.output.value}
+nodes:
+- name: print_env
+  type: python
+  source:
+    type: code
+    path: print_env.py
+  inputs:
+    key: ${inputs.key}
