+++ src/promptflow/tests/test_configs/flows/flow_with_connection_ref_environment_variables/inputs.jsonl
@@ -1,0 +1,6 @@
+{"text": "env1"}
+{"text": "env2"}
+{"text": "env3"}
+{"text": "env4"}
+{"text": "env5"}
+{"text": "env10"}
