+++ src/promptflow/tests/test_configs/flows/flow_with_connection_ref_environment_variables/print_env.py
@@ -1,0 +1,10 @@
+import os
+
+from promptflow import tool
+
+
+@tool
+def get_env_var(key: str):
+    print(os.environ.get(key))
+    # get from env var
+    return {"value": os.environ.get(key)}
